Updated Authorization in Controllers

+++ app/Http/Controllers/CommentsController.php
@@ -63,11 +63,15 @@ class CommentsController extends Controller
     public function update(Request $request, $id)
     {
         $comment = \App\Comment::find($id);
-        $comment->user_id = \Auth::user()->id;
-        $comment->post_id = $request->post_id;
-        $comment->comment_id = $request->comment_id;
-        $comment->content = $request->comment_content;
-        $comment->save();
+        if ($comment->user_id = \Auth::user()->id) {
+            $comment->post_id = $request->post_id;
+            $comment->comment_id = $request->comment_id;
+            $comment->content = $request->comment_content;
+            $comment->save();
+
+        } else { 
+            return response ("Unauthorized", 403);
+        }
 
         return $comment;
     }
@@ -81,7 +85,12 @@ class CommentsController extends Controller
     public function destroy($id)
     {
         $comment = \App\Comment::find($id);
-        $comment->delete();
+        if ($comment->user_id = \Auth::user()->id) {
+            $comment->delete();
+
+        } else { 
+            return response ("Unauthorized", 403);
+        }
 
         return $comment;
     }
